// admin/src/modulos/api/model/user.ts
/**
 * Receita Digital
 * Definição da API da Receita Digital 
 *
 * OpenAPI spec version: 1.0.6
 * Contact: fran.fig@gmail.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { UserLinks } from './userLinks';


export interface User {
    /**
     * ID único do usuário no sistema
     */
    id?: string;
    /**
     * Código CPF do usuário
     */
    cpf: string;
    /**
     * slug único do cliente (será o e-mail caso contrário)
     */
    username?: string;
    /**
     * e-mail do usuário no sistema
     */
    email: string;
    /**
     * Nome do cliente no sistema
     */
    nome?: string;
    /**
     * string MD5 da senha do cliente
     */
    password?: string;
    /**
     * celular do sistema para interação mobile
     */
    celular: string;
    /**
     * string com os escopos permitidos pelo usuário
     */
    scope?: string;
    links?: UserLinks;
    /**
     * User Status
     */
    userStatus?: number;
}