// admin/src/modulos/api/model/userLinks.ts
/**
 * Receita Digital
 * Definição da API da Receita Digital 
 *
 * OpenAPI spec version: 1.0.6
 * Contact: fran.fig@gmail.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { UserLinksFarmacia } from './userLinksFarmacia';


/**
 * ligação do usuário com os demais objetos do sistema, como atendente de farmacia, médico, paciente
 */
export interface UserLinks {
    /**
     * Código único do paciente do sistema
     */
    paciente?: string;
    /**
     * Código único do medico do sistema
     */
    medico?: string;
    farmacia?: Array<UserLinksFarmacia>;
}